rag-C1/requirements.txt
sentence-transformers>=3.0
rank-bm25>=0.2.2
pinecone>=5.0
pdfplumber>=0.11
torch>=2.0
pandas>=2.0
scikit-learn>=1.3
python-dotenv>=1.0
openai>=1.0
numpy>=1.24
